Proj/DanLuWangInterview/requirements.txt
asgiref==3.2.7
backcall==0.2.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
Django==2.2
django-cors-headers==3.4.0
djangorestframework==3.11.0
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
mysqlclient==1.4.6
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.2.4
